requirements.txt
Babel==2.6.0
Django==2.1.2
django-discover-runner==1.0
django-jwt-auth==0.0.2
django-phonenumber-field==2.0.1
django-rest-auth==0.9.3
django-role-permissions==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
Markdown==2.6.11
mongoengine==0.15.3
mysqlclient==1.3.13
phonenumbers==8.9.15
PyJWT==1.6.4
pymongo==3.7.1
pytz==2018.5
six==1.11.0
django-cors-headers==2.4.0
redis==2.10.6
requests==2.21.0
certifi==2018.11.29
urllib3==1.24.1
chardet==3.0.4 